# IGNITE-24732 Sql. Avoid expression lookup on every expression invocation