https://issues.apache.org/jira/browse/IGNITE-24732

Result of several queries from TPC-H suite on a single node cluster with scale factor 0.1:
```
Before
Benchmark          (fsync)  (queryId)  Mode  Cnt     Score    Error  Units
TpchBenchmark.run    false          1  avgt   20  1924.575 ±  3.575  ms/op
TpchBenchmark.run    false          2  avgt   20   256.358 ± 13.635  ms/op
TpchBenchmark.run    false          5  avgt   20   366.379 ±  0.677  ms/op
TpchBenchmark.run    false          7  avgt   20   548.897 ±  2.109  ms/op
TpchBenchmark.run    false          8  avgt   20   387.975 ±  3.322  ms/op
TpchBenchmark.run    false         22  avgt   20   417.330 ±  4.561  ms/op

After
Benchmark          (fsync)  (queryId)  Mode  Cnt     Score    Error  Units
TpchBenchmark.run    false          1  avgt   20   812.495 ±  3.205  ms/op
TpchBenchmark.run    false          2  avgt   20   257.707 ±  8.204  ms/op
TpchBenchmark.run    false          5  avgt   20   291.030 ±  0.732  ms/op
TpchBenchmark.run    false          7  avgt   20   253.482 ±  0.304  ms/op
TpchBenchmark.run    false          8  avgt   20   291.705 ±  0.796  ms/op
TpchBenchmark.run    false         22  avgt   20   362.490 ±  9.681  ms/op
```

----------
Thank you for submitting the pull request.

To streamline the review process of the patch and ensure better code quality
we ask both an author and a reviewer to verify the following:

### The Review Checklist
- [ ] **Formal criteria:** TC status, codestyle, mandatory documentation. Also make sure to complete the following:  
\- There is a single JIRA ticket related to the pull request.  
\- The web-link to the pull request is attached to the JIRA ticket.  
\- The JIRA ticket has the Patch Available state.  
\- The description of the JIRA ticket explains WHAT was made, WHY and HOW.  
\- The pull request title is treated as the final commit message. The following pattern must be used: IGNITE-XXXX Change summary where XXXX - number of JIRA issue.
- [ ] **Design:** new code conforms with the design principles of the components it is added to.
- [ ] **Patch quality:** patch cannot be split into smaller pieces, its size must be reasonable.
- [ ] **Code quality:** code is clean and readable, necessary developer documentation is added if needed.
- [ ] **Tests code quality:** test set covers positive/negative scenarios, happy/edge cases. Tests are effective in terms of execution time and resources.

### Notes
- [Apache Ignite Coding Guidelines](https://cwiki.apache.org/confluence/display/IGNITE/Java+Code+Style+Guide)